requirements.txt
absl-py==1.4.0
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Cython==3.0.0
exceptiongroup==1.1.2
fastapi==0.101.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
joblib==1.3.1
keras==2.13.1
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
patsy==0.5.3
pmdarima==2.0.3
protobuf==4.24.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic_core==2.4.0
PyMeeus==0.5.12
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
setuptools-git==1.2
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
statsmodels==0.14.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.5
uvicorn==0.23.2
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
